+++ Runtime/Framework/IoC.cs
@@ -8,9 +8,19 @@ namespace DT.UniStart {
   /// </summary>
   public interface IIoCC {
     /// <summary>
+    /// Register a type with an existing instance and a key.
+    /// </summary>
+    T Add<T>(object key, T instance);
+
+    /// <summary>
     /// Register a type with an existing instance.
     /// </summary>
     T Add<T>(T instance);
+
+    /// <summary>
+    /// Register a type with a key, and auto create an instance.
+    /// </summary>
+    T Add<T>(object key) where T : new();
 
     /// <summary>
     /// Register a type and auto create an instance.
@@ -18,9 +28,20 @@ namespace DT.UniStart {
     T Add<T>() where T : new();
 
     /// <summary>
+    /// Get the instance of a type by key.
+    /// </summary>
+    T Get<T>(object key);
+
+    /// <summary>
     /// Get the instance of a type.
     /// </summary>
     T Get<T>();
+
+    /// <summary>
+    /// Try to get the instance of a type with a key.
+    /// If the type is not registered, return `default(T)`.
+    /// </summary>
+    T TryGet<T>(object key);
 
     /// <summary>
     /// Try to get the instance of a type.
@@ -33,18 +54,32 @@ namespace DT.UniStart {
   /// IoC Container.
   /// </summary>
   public class IoCC : IIoCC {
-    Dictionary<Type, object> dict;
+    Dictionary<object, object> dict;
 
     public IoCC() {
-      this.dict = new Dictionary<Type, object>();
+      this.dict = new Dictionary<object, object>();
+    }
+
+    /// <summary>
+    /// Register a type with an existing instance and a key.
+    /// </summary>
+    public T Add<T>(object key, T instance) {
+      this.dict.Add(key, instance);
+      return instance;
     }
 
     /// <summary>
     /// Register a type with an existing instance.
     /// </summary>
     public T Add<T>(T instance) {
-      this.dict.Add(typeof(T), instance);
-      return instance;
+      return this.Add(typeof(T), instance);
+    }
+
+    /// <summary>
+    /// Register a type with a key, and auto create an instance.
+    /// </summary>
+    public T Add<T>(object key) where T : new() {
+      return this.Add<T>(key, new T());
     }
 
     /// <summary>
@@ -55,10 +90,29 @@ namespace DT.UniStart {
     }
 
     /// <summary>
+    /// Get the instance of a type by key.
+    /// </summary>
+    public T Get<T>(object key) {
+      return (T)this.dict[key];
+    }
+
+    /// <summary>
     /// Get the instance of a type.
     /// </summary>
     public T Get<T>() {
-      return (T)this.dict[typeof(T)];
+      return this.Get<T>(typeof(T));
+    }
+
+    /// <summary>
+    /// Try to get the instance of a type with a key.
+    /// If the type is not registered, return `default(T)`.
+    /// </summary>
+    public T TryGet<T>(object key) {
+      if (this.dict.TryGetValue(key, out object value)) {
+        return (T)value;
+      } else {
+        return default(T);
+      }
     }
 
     /// <summary>
@@ -66,11 +120,7 @@ namespace DT.UniStart {
     /// If the type is not registered, return `default(T)`.
     /// </summary>
     public T TryGet<T>() {
-      if (this.dict.TryGetValue(typeof(T), out object value)) {
-        return (T)value;
-      } else {
-        return default(T);
-      }
+      return this.TryGet<T>(typeof(T));
     }
   }
 }